Add Generic Buffer Sound source and make system more scalable
Previously the connection between C++ and Kotlin for sound source identification / mutation was determined by an enum value. This isn't scalable since the Generic sound source can be many different sounds but is a single enum type. 

Instead, we've created a unique identifier as a string for each possible sound source the app can have at a time. Currently, I've written this as a hardcoded list, however in the future this could be configured in a database. 

SoundSource creation in C++ is now done through a factory object with 2 overloads. The sound mixer stores a map of the unique id to SoundSource. 

Furthermore, improved the return types of certain functions to return errors if they happen (via nullptrs, id not in map, etc).