Currently Playing Samples (#7)
## Task

Provide some awareness of the currently playing samples in the app. Provide a player ribbon with a basic UX for controls. UX to be improved.

Also handle placement of Audio Player clean up - this is temporary, eventually this needs to be refactored into a Frontend Service so it will run after application destruction.

## Solution
code the stuff up

## Testing
No tests were created - manual testing only